% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.subtaction.byhand.r
\name{baseline.subtraction.byhand}
\alias{baseline.subtraction.byhand}
\title{Subtracts the baseline from the function given two endpoints}
\usage{
baseline.subtraction.byhand(x, lwr.temp, upr.temp, plot.on = TRUE)
}
\arguments{
\item{x}{thermogram data}

\item{lwr.temp}{The lower endpoint}

\item{upr.temp}{The upper endpoint}

\item{plot.on}{logical: should a graph of the baseline subtraction be produced?}
}
\value{
Data frame of temperature and dCp with the baseline subtracted
}
\description{
Given two endpoints and the thermogram data, the function will connect the two points with a line, then subtract the line from the thermogram.
}
