% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.sample.interpolate.r
\name{final.sample.interpolate}
\alias{final.sample.interpolate}
\title{Interpolates each sample on a fixed grid of temperature}
\usage{
final.sample.interpolate(x, grid.temp, plot.on = TRUE)
}
\arguments{
\item{x}{Baseline subtracted thermogram data}

\item{grid.temp}{The grid of temperatures the sample is interpolated onto}

\item{plot.on}{logical: should the output be graphed?}
}
\value{
Data frame of temperature and dCp
}
\description{
Fits a spline to the baseline subtracted data and interpolates the thermogram on the given temperature grid.
}
