% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrineWorking.R
\docType{data}
\name{UrineWorking}
\alias{UrineWorking}
\title{Clean Urine Thermogram Data}
\format{
A data frame with 3118189 observations with 3 columns.
\describe{
\item{SampleID}{A unique identifier for each sample and iteration}
\item{Temperature}{The temperature measured in degrees celsius}
\item{dCp}{The dCp at the given temperature}
}
}
\usage{
UrineWorking
}
\description{
A data frame containing 393 urine thermograms in "long" format. Each sample has two iterations "a" and "b" for a total of 781 thermograms.
}
\keyword{datasets}
